requirements.txt
azure-core==1.34.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
Flask==3.1.1
Flask-Login==0.6.3
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0

pyodbc~=5.2.0
python-dotenv~=1.1.1